Move more consts
