+++ packages/back-end/src/util/migrations.ts
@@ -9,6 +9,7 @@ import { RESERVED_ROLE_IDS, getDefaultRole } from "shared/permissions";
 import { omit } from "lodash";
 import { SavedGroupInterface } from "shared/src/types";
 import { v4 as uuidv4 } from "uuid";
+import { accountFeatures } from "shared/enterprise";
 import {
   ExperimentReportArgs,
   ExperimentReportInterface,
@@ -42,7 +43,7 @@ import {
 } from "back-end/types/experiment-snapshot";
 import { getEnvironments } from "back-end/src/services/organizations";
 import { LegacySavedGroupInterface } from "back-end/types/saved-group";
-import { accountFeatures, getAccountPlan } from "back-end/src/enterprise";
+import { getAccountPlan } from "back-end/src/enterprise";
 import { DEFAULT_CONVERSION_WINDOW_HOURS } from "./secrets";
 
 function roundVariationWeight(num: number): number {
